refactor: move watchman watch and caching into runner/pkg/
