+++ common/cache/BUILD.bazel
@@ -7,14 +7,12 @@ go_library(
         "configurer.go",
         "disk.go",
         "noop.go",
-        "watchman.go",
     ],
     importpath = "github.com/aspect-build/aspect-gazelle/common/cache",
     visibility = ["//visibility:public"],
     deps = [
         "//common/buildinfo",
         "//common/logger",
-        "//common/watch",
         "@gazelle//config",
         "@gazelle//language",
         "@gazelle//rule",
