--- common/watch/BUILD.bazel
@@ -1,19 +1,0 @@
-load("@rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "watch",
-    srcs = ["watch.go"],
-    importpath = "github.com/aspect-build/aspect-gazelle/common/watch",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//common/bazel",
-        "//common/socket",
-    ],
-)
-
-go_test(
-    name = "watch_test",
-    srcs = ["watch_test.go"],
-    embed = [":watch"],
-    tags = ["manual"],  # TODO: we need watchman in our runners.
-)
